pyproject.toml
[project]
name = "mcp-modules"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.17.0",
    "py-solc-x>=2.0.4",
    "pytest>=8.4.2",
    "slither-analyzer>=0.11.3",
]

[tool.setuptools]
packages = ["mcp_modules"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
